+++ DeslSearch/DeslSearch/cipher.cpp
@@ -8,7 +8,7 @@ int PTable[32] = { 15, 6, 19, 20, 28, 11, 27, 16,
 };
 
 //扩展置换E中调用的函数，从input的index位置开始取6个比特
-void pick6(u8* output, u32 input, int index){
+void pick6(u16* output, u32 input, int index){
 	if (index > 26){
 		int carry = 32 - index;
 		*output = ((input << (index - 26)) ^ (input >> (58 - index))) & 0x3f;
@@ -19,7 +19,7 @@ void pick6(u8* output, u32 input, int index){
 }
 
 //扩展置换E
-void Expansion(u8* output, u32 input){
+void Expansion(u16* output, u32 input){
 	for (int i = 0; i < 8; i++){
 		pick6(output + i, input, ((4 * i - 1) % 32+32)%32);
 	}
@@ -37,7 +37,7 @@ u32 Permutation(u32 x){
 }
 
 //8个4比特数组转换成32比特串
-u32 SboxOutput2word(u8* input){
+u32 SboxOutput2word(u16* input){
 	u32 output=0;
 	for(int i=0;i<7;i++){
 		output+=*(input+i);
@@ -50,7 +50,7 @@ u32 SboxOutput2word(u8* input){
 //输入是S盒输出，是8个4比特数组
 //输出是E的输出，是8个6比特数组
 //实现线性层
-void PE(u8* output, u8* input){
+void PE(u16* output, u16* input){
 	u32 temp=SboxOutput2word(input);
 	temp=Permutation(temp);
 	Expansion(output,temp);
